--- src/enip/cip/index.js
@@ -1,7 +1,0 @@
-const MessageRouter = require("./message-router");
-const DataTypes = require("./data-types");
-const EPATH = require("./epath");
-const UnconnectedSend = require("./unconnected-send");
-const ConnectionManager = require("./connection-manager");
-
-module.exports = { MessageRouter, DataTypes, EPATH, UnconnectedSend, ConnectionManager };
